Extend last_modified_by max_length to 150
Previus the length for last_modified_by is 10 for Feature and HistoricalFeature, and 20 for Transaction. It crashes when a user with a username that longer than that updates a feature or transaction. We extend the max_length of the fields to the same length as the username field.

no refs