Refactor filtering logic of protected data
The previous solution did not work as expected and did not take into account
several models and how they should be filtered. It was also unnessecary complex
in that it used multiple managers and queryset classes.
For instance, with the old filtering system, when a FeatureClass was set to non open data
observations of features that belonged to the feature class could still be fetched.
Another issue that the old solution had was that there was no way to just pass in a
authorization class and get back the content for that class, but a specific function
had to be called.

The new solution is fully backwards compatible with the old APIs and all functions
that existed before still do so. The model manager and queryset has now been
made more generic so that they can be used in several different models that has
a connection to a model that has either a protection level or open data restrictions.
On top of that, multiple filters can now be applied to a query in order to combine both
open data filtering and protection level filters.