Clean up the open feature classes in migrations
Instead of using feature classes, we are now using `open_data` field to
indicate if a feature can be published as open data. The open feature
classes are changing over the time, and we should not rely on the
feature classes to determine weather a feature is open data or not.

Refs: #285